% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colf_nlxb_methods.R
\name{fitted.colf_nlxb}
\alias{fitted.colf_nlxb}
\title{Fitted values for colf_nlxb}
\usage{
\method{fitted}{colf_nlxb}(object, ...)

\method{fitted}{colf_nlxb}(object, ...)
}
\arguments{
\item{object}{A colf_nlxb object i.e. the result of running \code{colf_nlxb}}

\item{...}{Currently not used}

\item{object}{A colf_nlxb object i.e. the result of running \code{colf_nlxb}}

\item{...}{Currently not used}
}
\value{
A vector with the fitted values

A vector with the coefficients
}
\description{
Fitted values for colf_nlxb

Coefficients for colf_nlxb
}
\examples{
mymod <- colf_nlxb(mpg ~ hp + cyl, mtcars)

#fitted values
fitted(mymod)

mymod <- colf_nlxb(mpg ~ hp + cyl, mtcars)

#coefficients
coef(mymod)

}

