% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colf_nlxb_methods.R
\name{print.colf_nlxb}
\alias{print.colf_nlxb}
\title{colf_nlxb Print method}
\usage{
\method{print}{colf_nlxb}(x, ...)
}
\arguments{
\item{x}{A colf_nlxb object i.e. the result of running \code{colf_nlxb}}

\item{...}{Currently not used}
}
\value{
Printing the colf_nlxb object
}
\description{
colf_nlxb Print method
}
\examples{
mymod <- colf_nlxb(mpg ~ hp + cyl, mtcars)

#print
print(mymod)

}

