% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colf_nlxb_methods.R
\name{residuals.colf_nlxb}
\alias{residuals.colf_nlxb}
\title{Residuals for colf_nlxb}
\usage{
\method{residuals}{colf_nlxb}(object, ...)
}
\arguments{
\item{object}{A colf_nlxb object i.e. the result of running \code{colf_nlxb}}

\item{...}{Currently not used}
}
\value{
A vector with the residuals
}
\description{
Residuals for colf_nlxb
}
\examples{
mymod <- colf_nlxb(mpg ~ hp + cyl, mtcars)

#residuals
residuals(mymod)
resid(mymod)

}

