% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colf_nlxb_methods.R
\name{summary.colf_nlxb}
\alias{summary.colf_nlxb}
\title{colf_nlxb Summary}
\usage{
\method{summary}{colf_nlxb}(object, ...)
}
\arguments{
\item{object}{A colf_nlxb object i.e. the result of running \code{colf_nlxb}}

\item{...}{Currently not used}
}
\value{
The summary of the model
}
\description{
colf_nlxb Summary
}
\examples{
mymod <- colf_nlxb(mpg ~ hp + cyl, mtcars)

#summary
summary(mymod)

}

