% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colf_nlxb.R
\name{colf_nlxb}
\alias{colf_nlxb}
\title{Nash Variant of the Marquardt algorithm on a linear objective function}
\usage{
colf_nlxb(formula, data, start = NULL, trace = FALSE, lower = -Inf,
  upper = Inf, masked = NULL, control)
}
\arguments{
\item{formula}{The formula. This has the same syntax as formula in \code{lm}. Look at details.}

\item{data}{A data frame containing the data of the variables in the formula.}

\item{start}{A named parameter list. Check details.}

\item{trace}{Logical TRUE if we want intermediate progress to be reported. Default is FALSE.}

\item{lower}{Lower bounds on the parameters. If a single number, this will be applied to all 
parameters. Default -Inf.}

\item{upper}{Upper bounds on the parameters. If a single number, this will be applied to all 
parameters. Default Inf.}

\item{masked}{Character vector of quoted parameter names. These parameters will NOT be altered 
by the algorithm.}

\item{control}{A list of controls for the algorithm. These are:}
}
\value{
Same as nlxb
}
\description{
Non linear least squares solution via qr linear solver on a linear objective function.
}
\details{
\code{colf_nlxb} uses Nash's (Nash, 1979) variant of the Marquardt algorithm, in an attempt to
find the minimum of the residual sum of squares.
}

